+++ src/iter.rs
@@ -2,10 +2,7 @@
 # Dowser: Dowser
 */
 
-use crate::{
-	NoHashState,
-	NoHashU64,
-};
+use crate::NoHashState;
 
 #[cfg(feature = "parking_lot_mutex")]
 use parking_lot::Mutex;
@@ -17,6 +14,9 @@ use rayon::iter::{
 };
 use std::{
 	collections::HashSet,
+	fs::DirEntry,
+	hash::Hasher,
+	num::NonZeroUsize,
 	os::unix::fs::MetadataExt,
 	path::{
 		Path,
@@ -39,7 +39,77 @@ macro_rules! mutex { ($var:expr) => ($var.lock().unwrap_or_else(std::sync::Poiso
 
 
 
-#[derive(Debug)]
+/// # Resolved Path.
+///
+/// This tuple represents the hash, whether or not it is a directory, and the
+/// canonical path.
+type Resolved = (u64, bool, PathBuf);
+
+
+
+#[derive(Debug, Clone, Copy)]
+/// # Directory Concurrency.
+///
+/// This enum determines if and how many directories [`Dowser`] should try to
+/// read in parallel, which is configured via [`Dowser::with_dir_concurrency`].
+///
+/// The default is [`DirConcurrency::Sane`], which caps the maximum number of
+/// concurrent directory reads to `(CPU/Threads - 1).min(8)`. This is a good
+/// middle ground between [`DirConcurrency::Single`] and [`DirConcurrency::Max`].
+///
+/// If you anticipate there being very few directories, [`DirConcurrency::Single`]
+/// could actually be faster.
+///
+/// Conversely, if you expect _a lot_ of directories, [`DirConcurrency::Max`] is
+/// likely the best strategy. However **be careful** with this choice. If the user's
+/// `ulimit` is set too low, paths might be intermittently skipped due to
+/// unreadability.
+pub enum DirConcurrency {
+	/// # One at a Time.
+	Single,
+
+	/// # A Probably Okay Default.
+	Sane,
+
+	/// # READ EVERYTHING ALWAYS.
+	Max,
+
+	/// # A Custom Value.
+	Other(NonZeroUsize),
+}
+
+impl Default for DirConcurrency {
+	fn default() -> Self { Self::Sane }
+}
+
+impl From<usize> for DirConcurrency {
+	fn from(src: usize) -> Self {
+		match src {
+			0 => Self::Max,
+			1 => Self::Single,
+			// Safety: zero is checked above.
+			n => Self::Other(unsafe { NonZeroUsize::new_unchecked(n) }),
+		}
+	}
+}
+
+impl From<DirConcurrency> for usize {
+	fn from(src: DirConcurrency) -> Self {
+		match src {
+			DirConcurrency::Single => 1,
+			DirConcurrency::Sane => match rayon::current_num_threads() {
+				0..=2 => 1,
+				n => Self::min(n - 1, 8),
+			},
+			DirConcurrency::Max => 0,
+			DirConcurrency::Other(n) => n.get(),
+		}
+	}
+}
+
+
+
+#[derive(Debug, Clone)]
 /// # Dowser.
 ///
 /// `Dowser` is a very simple recursive file iterator with parallelized
@@ -89,7 +159,7 @@ macro_rules! mutex { ($var:expr) => ($var.lock().unwrap_or_else(std::sync::Poiso
 pub struct Dowser {
 	files: Vec<PathBuf>,
 	dirs: Vec<PathBuf>,
-	threads: usize,
+	dir_concurrency: usize,
 	seen: HashSet<u64, NoHashState>,
 }
 
@@ -98,7 +168,7 @@ impl Default for Dowser {
 		Self {
 			files: Vec::with_capacity(8),
 			dirs: Vec::with_capacity(8),
-			threads: dowser_threads(),
+			dir_concurrency: usize::from(DirConcurrency::Sane),
 			seen: HashSet::with_capacity_and_hasher(4096, NoHashState::default()),
 		}
 	}
@@ -106,24 +176,16 @@ impl Default for Dowser {
 
 impl From<PathBuf> for Dowser {
 	fn from(src: PathBuf) -> Self {
-		if let Some((h, is_dir, p)) = resolve_path(src, false) {
-			let mut seen = HashSet::with_capacity_and_hasher(4096, NoHashState::default());
-			seen.insert(h);
+		let mut out = Self::default();
 
-			let mut dirs = Vec::with_capacity(8);
-			let mut files = Vec::with_capacity(8);
-
-			if is_dir { dirs.push(p); }
-			else { files.push(p); }
-
-			Self {
-				files,
-				dirs,
-				threads: dowser_threads(),
-				seen,
+		if let Some((h, is_dir, p)) = resolve_path(src) {
+			if out.seen.insert(h) {
+				if is_dir { out.dirs.push(p); }
+				else { out.files.push(p); }
 			}
 		}
-		else { Self::default() }
+
+		out
 	}
 }
 
@@ -137,45 +199,31 @@ impl From<&PathBuf> for Dowser {
 
 impl From<&[PathBuf]> for Dowser {
 	fn from(src: &[PathBuf]) -> Self {
-		let mut seen = HashSet::with_capacity_and_hasher(4096, NoHashState::default());
-		let mut dirs = Vec::with_capacity(8);
-		let mut files = Vec::with_capacity(8);
+		let mut out = Self::default();
 
-		for (h, is_dir, p) in src.iter().filter_map(|p| resolve_path(p.clone(), false)) {
-			if seen.insert(h) {
-				if is_dir { dirs.push(p); }
-				else { files.push(p); }
+		for (h, is_dir, p) in src.iter().filter_map(|p| resolve_path(p.clone())) {
+			if out.seen.insert(h) {
+				if is_dir { out.dirs.push(p); }
+				else { out.files.push(p); }
 			}
 		}
 
-		Self {
-			files,
-			dirs,
-			threads: dowser_threads(),
-			seen,
-		}
+		out
 	}
 }
 
 impl From<Vec<PathBuf>> for Dowser {
 	fn from(src: Vec<PathBuf>) -> Self {
-		let mut seen = HashSet::with_capacity_and_hasher(4096, NoHashState::default());
-		let mut dirs = Vec::with_capacity(8);
-		let mut files = Vec::with_capacity(8);
+		let mut out = Self::default();
 
-		for (h, is_dir, p) in src.into_iter().filter_map(|p| resolve_path(p, false)) {
-			if seen.insert(h) {
-				if is_dir { dirs.push(p); }
-				else { files.push(p); }
+		for (h, is_dir, p) in src.into_iter().filter_map(resolve_path) {
+			if out.seen.insert(h) {
+				if is_dir { out.dirs.push(p); }
+				else { out.files.push(p); }
 			}
 		}
 
-		Self {
-			files,
-			dirs,
-			threads: dowser_threads(),
-			seen,
-		}
+		out
 	}
 }
 
@@ -188,73 +236,56 @@ impl Iterator for Dowser {
 			if let Some(p) = self.files.pop() {
 				return Some(p);
 			}
-			// Read some directories.
-			match self.dirs.len().min(self.threads) {
-				0 => break,
-				1 => { self.crawl(); },
-				n => { self.crawl_n(n); },
+
+			// Read some directories maybe!
+			let len = self.dirs.len();
+			if len == 0 { break; }
+			if self.dir_concurrency == 1 || len == 1 {
+				if let Ok(rd) = std::fs::read_dir(self.dirs.remove(len - 1)) {
+					for (h, is_dir, p) in rd.filter_map(resolve_entry) {
+						if self.seen.insert(h) {
+							if is_dir { self.dirs.push(p); }
+							else { self.files.push(p); }
+						}
+					}
+				}
+			}
+			else {
+				let idx =
+					if self.dir_concurrency == 0 { 0 }
+					else { len.saturating_sub(self.dir_concurrency) };
+
+				let new = self.dirs.split_off(idx);
+				let seen = Mutex::new(&mut self.seen);
+				let f = Mutex::new(&mut self.files);
+				let d = Mutex::new(&mut self.dirs);
+
+				new.into_par_iter()
+					.filter_map(|p| std::fs::read_dir(p).ok())
+					.flat_map(ParallelBridge::par_bridge)
+					.for_each(|e| if let Some((h, is_dir, p)) = resolve_entry(e) {
+						if mutex!(seen).insert(h) {
+							if is_dir { mutex!(d).push(p); }
+							else { mutex!(f).push(p); }
+						}
+					});
 			}
 		}
 
 		None
 	}
 
+	/// # Size Hints.
+	///
+	/// This iterator has an unknown size until the final directory has been
+	/// read.
 	fn size_hint(&self) -> (usize, Option<usize>) {
-		let dirs = self.dirs.len();
-		let files = self.files.len();
+		let lower = self.files.len();
+		let upper =
+			if self.dirs.is_empty() { Some(lower) }
+			else { None };
 
-		if 0 == dirs { (files, Some(files)) }
-		else { (files + dirs, None) }
-	}
-}
-
-impl Dowser {
-	/// # Crunch One.
-	///
-	/// This reads the final directory in the queue when there is only one.
-	fn crawl(&mut self) {
-		if let Some(p) = self.dirs.pop() {
-			if let Ok(rd) = std::fs::read_dir(p) {
-				let seen = Mutex::new(&mut self.seen);
-				let f = Mutex::new(&mut self.files);
-				let d = Mutex::new(&mut self.dirs);
-
-				rd.par_bridge()
-					.for_each(|e| if let Ok(e) = e {
-						if let Some((h, is_dir, p)) = resolve_path(e.path(), true) {
-							if mutex!(seen).insert(h) {
-								if is_dir { mutex!(d).push(p); }
-								else { mutex!(f).push(p); }
-							}
-						}
-					});
-			}
-		}
-	}
-
-	/// # Crunch Many.
-	///
-	/// This reads multiple directories — but maybe not _all_ — in the queue,
-	/// using multiple threads to spread the effort out a bit.
-	fn crawl_n(&mut self, n: usize) {
-		// Split off so we can write right back to self.dirs during iteration.
-		let new = self.dirs.split_off(self.dirs.len() - n);
-
-		let seen = Mutex::new(&mut self.seen);
-		let f = Mutex::new(&mut self.files);
-		let d = Mutex::new(&mut self.dirs);
-
-		new.into_par_iter()
-			.filter_map(|p| std::fs::read_dir(p).ok())
-			.flat_map(ParallelBridge::par_bridge)
-			.for_each(|e| if let Ok(e) = e {
-				if let Some((h, is_dir, p)) = resolve_path(e.path(), true) {
-					if mutex!(seen).insert(h) {
-						if is_dir { mutex!(d).push(p); }
-						else { mutex!(f).push(p); }
-					}
-				}
-			});
+		(lower, upper)
 	}
 }
 
@@ -307,13 +338,26 @@ impl Dowser {
 	/// ```
 	pub fn with_path<P>(mut self, path: P) -> Self
 	where P: AsRef<Path> {
-		if let Some((h, is_dir, p)) = resolve_path(PathBuf::from(path.as_ref()), false) {
+		if let Some((h, is_dir, p)) = resolve_path(path.as_ref().to_path_buf()) {
 			if self.seen.insert(h) {
 				if is_dir { self.dirs.push(p); }
 				else { self.files.push(p); }
 			}
 		}
 
+		self
+	}
+
+	#[must_use]
+	/// # With Directory Concurrency.
+	///
+	/// By default, [`Dowser`] processes directories in parallel, but only so
+	/// many at a time. This tends to provide a small performance boost for
+	/// most searches, but may not always be the best strategy.
+	///
+	/// See [`DirConcurrency`] for more information.
+	pub fn with_dir_concurrency(mut self, val: DirConcurrency) -> Self {
+		self.dir_concurrency = usize::from(val);
 		self
 	}
 }
@@ -381,19 +425,105 @@ impl Dowser {
 	}
 }
 
+impl Dowser {
+	#[must_use]
+	/// # Consume Into Vec (Filtered).
+	///
+	/// This method is an optimized alternative to running
+	/// `Dowser.iter().filter(…).collect::<Vec<PathBuf>>()`.
+	///
+	/// It yields the same results as the above, but makes fewer allocations
+	/// along the way and applies your filter callback in parallel.
+	///
+	/// ## Examples
+	///
+	/// ```no_run
+	/// use dowser::{Dowser, Extension};
+	/// use std::path::PathBuf;
+	///
+	/// const GZ: Extension = Extension::new2(*b"gz");
+	///
+	/// // The iterator way.
+	/// let files: Vec<PathBuf> = Dowser::default()
+	///     .with_path("/usr/share")
+	///     .filter(|p| Some(GZ) == Extension::try_from2(p))
+	///     .collect();
+	///
+	/// // The optimized way.
+	/// let files: Vec<PathBuf> = Dowser::default()
+	///     .with_path("/usr/share")
+	///     .into_vec(|p| Some(GZ) == Extension::try_from2(p));
+	/// ```
+	pub fn into_vec<F>(self, cb: F) -> Vec<PathBuf>
+	where F: Fn(&Path) -> bool + Sync + Send {
+		let Self { mut files, mut dirs, dir_concurrency, mut seen } = self;
 
+		// We wouldn't have had a chance to filter these yet.
+		if ! files.is_empty() {
+			files.retain(|p| cb(p));
+		}
 
-#[inline]
-/// # Thread Count.
-///
-/// This returns the ideal number of threads to use when crawling directories.
-/// To help with `ulimit` difficulties, this is either one less than what Rayon
-/// would normally use, or 8, whichever is lower.
-fn dowser_threads() -> usize {
-	match rayon::current_num_threads() {
-		0..=2 => 1,
-		n => usize::min(n - 1, 8),
+		// Consume!
+		{
+			let s = Mutex::new(&mut seen);
+			let f = Mutex::new(&mut files);
+
+			loop {
+				let len = dirs.len();
+				if len == 0 { break; }
+				let idx =
+					if dir_concurrency == 0 { 0 }
+					else { len.saturating_sub(dir_concurrency) };
+
+				let new = dirs.split_off(idx);
+				let d = Mutex::new(&mut dirs);
+
+				new.into_par_iter()
+					.filter_map(|p| std::fs::read_dir(p).ok())
+					.flat_map(ParallelBridge::par_bridge)
+					.for_each(|e| if let Some((h, is_dir, p)) = resolve_entry(e) {
+						if mutex!(s).insert(h) {
+							if is_dir { mutex!(d).push(p); }
+							else if cb(&p) { mutex!(f).push(p); }
+						}
+					});
+			}
+		}
+
+		// Done!
+		files
 	}
+}
+
+
+
+/// # Hash Path.
+fn hash_path(dev: u64, ino: u64) -> u64 {
+	let mut hasher = ahash::AHasher::new_with_keys(1319, 2371);
+	hasher.write_u64(dev);
+	hasher.write_u64(ino);
+	hasher.finish()
+}
+
+/// # Resolve Entry Result.
+///
+/// This is like `resolve_path`, except it assumes that any non-symlink entry
+/// path is canonical, because its parent path was canonical.
+///
+/// Symlinks get passed to `resolve_path`, ensuring they're fully
+/// canonicalized.
+fn resolve_entry(e: Result<DirEntry, std::io::Error>) -> Option<Resolved> {
+	let e = e.ok()?;
+
+	// If this is a symlink, we have to follow it.
+	if e.file_type().map_or(true, |ft| ft.is_symlink()) {
+		return resolve_path(e.path());
+	}
+
+	let meta = e.metadata().ok()?;
+	let path = e.path();
+	let hash: u64 = hash_path(meta.dev(), meta.ino());
+	Some((hash, meta.is_dir(), path))
 }
 
 /// # Resolve Path.
@@ -402,27 +532,10 @@ fn dowser_threads() -> usize {
 /// * A unique hash derived from the path's device and inode.
 /// * A bool indicating whether or not the path is a directory.
 /// * The canonicalized path.
-///
-/// As [`std::fs::canonicalize`] is an expensive operation, this method allows
-/// a "trusted" bypass, which will only canonicalize the path if it is a
-/// symlink.
-///
-/// The trusted mode is only appropriate in cases like `ReadDir` where the
-/// directory seed was canonicalized. The idea is that since `DirEntry` paths
-/// are joined to the seed, they'll be canonical so long as the seed was,
-/// except in cases of symlinks.
-fn resolve_path(path: PathBuf, trusted: bool) -> Option<(u64, bool, PathBuf)> {
-	if trusted {
-		let meta = std::fs::symlink_metadata(&path).ok()?;
-		if ! meta.file_type().is_symlink() {
-			let hash: u64 = NoHashU64::hash_path(meta.dev(), meta.ino());
-			return Some((hash, meta.is_dir(), path));
-		}
-	}
-
+fn resolve_path(path: PathBuf) -> Option<Resolved> {
 	let path = std::fs::canonicalize(path).ok()?;
 	let meta = std::fs::metadata(&path).ok()?;
-	let hash: u64 = NoHashU64::hash_path(meta.dev(), meta.ino());
+	let hash: u64 = hash_path(meta.dev(), meta.ino());
 	Some((hash, meta.is_dir(), path))
 }
 
@@ -430,17 +543,17 @@ fn resolve_path(path: PathBuf, trusted: bool) -> Option<(u64, bool, PathBuf)> {
 ///
 /// This is identical to `resolve_path`, except it only returns the hash. It
 /// is used by [`Dowser::without_paths`] and [`Dowser::without_path`], which
-/// don't actually need anything more.
+/// don't need the rest.
 fn resolve_path_hash(path: &Path) -> Option<u64> {
 	if let Ok(meta) = std::fs::symlink_metadata(&path) {
-		if ! meta.file_type().is_symlink() {
-			return Some(NoHashU64::hash_path(meta.dev(), meta.ino()));
+		if ! meta.is_symlink() {
+			return Some(hash_path(meta.dev(), meta.ino()));
 		}
 	}
 
 	let path = std::fs::canonicalize(path).ok()?;
 	let meta = std::fs::metadata(&path).ok()?;
-	Some(NoHashU64::hash_path(meta.dev(), meta.ino()))
+	Some(hash_path(meta.dev(), meta.ino()))
 }
 
 
@@ -480,6 +593,15 @@ mod tests {
 		let test_dir = std::fs::canonicalize("./tests/links")
 			.expect("Missing dowser link directory.");
 
+		// Make sure symlinks are detected.
+		let links = std::fs::read_dir(&test_dir)
+			.expect("Missing dowser link directory.")
+			.filter_map(Result::ok)
+			.filter_map(|e| e.file_type().ok())
+			.filter(|t| t.is_symlink())
+			.count();
+		assert_eq!(links, 1, "Wrong symlink count!");
+
 		let raw = vec![
 			test_dir.join("01"),
 			test_dir.join("02"),
@@ -491,9 +613,10 @@ mod tests {
 			test_dir.join("06/08"),
 			test_dir.join("06/09"),
 			test_dir.join("06/10"), // Sym to one.
+			test_dir.join("06/11"), // Sym to seven to six.
 		];
 
-		let canon = {
+		let mut canon = {
 			let mut tmp: Vec<PathBuf> = raw.iter()
 				.filter_map(|x| std::fs::canonicalize(x).ok())
 				.collect();
@@ -503,14 +626,15 @@ mod tests {
 		};
 
 		// There should be two fewer entries as two are symlinks.
-		assert_eq!(raw.len(), 10);
+		assert_eq!(raw.len(), 11);
 		assert_eq!(canon.len(), 8, "{:?}", canon);
 		assert!(! canon.contains(&raw[6]));
 		assert!(! canon.contains(&raw[9]));
+		assert!(! canon.contains(&raw[10]));
 
 		let trusting = {
 			let mut tmp: Vec<PathBuf> = raw.iter()
-				.filter_map(|x| resolve_path(x.clone(), true).map(|(_, _, p)| p))
+				.filter_map(|x| resolve_path(x.clone()).map(|(_, _, p)| p))
 				.collect();
 			tmp.sort();
 			tmp.dedup();
@@ -518,5 +642,17 @@ mod tests {
 		};
 
 		assert_eq!(trusting, canon);
+
+		// Now let's make sure Dowser does the same thing, albeit just the
+		// files.
+		canon.retain(|p| p.is_file());
+
+		let mut itered: Vec<PathBuf> = Dowser::from(test_dir.as_path()).collect();
+		itered.sort();
+		assert_eq!(canon, itered);
+
+		itered = Dowser::from(test_dir.as_path()).into_vec(|_| true);
+		itered.sort();
+		assert_eq!(canon, itered);
 	}
 }