+++ src/lib.rs
@@ -87,8 +87,6 @@ let files: Vec::<PathBuf> = Dowser::default()
 ```
 */
 
-#![forbid(unsafe_code)]
-
 #![warn(clippy::filetype_is_file)]
 #![warn(clippy::integer_division)]
 #![warn(clippy::needless_borrow)]
@@ -120,11 +118,11 @@ pub mod utility;
 
 
 
-pub use iter::Dowser;
+pub use iter::{
+	DirConcurrency,
+	Dowser,
+};
 pub use ext::Extension;
 
 #[doc(hidden)]
-pub(crate) use hash::{
-	NoHashU64,
-	NoHashState,
-};
+pub(crate) use hash::NoHashState;