+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 
 
+## [0.4.3](https://github.com/Blobfolio/dowser/releases/tag/v0.4.3) - 2022-03-26
+
+### Added
+
+* impl `Clone` for `Dowser`
+* `Dowser::into_vec`
+* `Dowser::with_dir_concurrency`
+
+
+
 ## [0.4.2](https://github.com/Blobfolio/dowser/releases/tag/v0.4.2) - 2022-03-08
 
 ### Changed